Qualify function calls when the compiler can optimize them
Some built-in functions can be turned into specific opcodes at compile
time when the compiler is sure that they will be used (i.e. no namespace
fallback is needed). This can have a major performance impact when they
appear in hot paths.
In the profiling, is_null was one of the most expansive function due to
the huge number of times it was used. And it is one of the functions
which is optimized.

This change has been done using the php-cs-fixer tool, which maintains a
list of functions which benefit from such optimization.