Fix the handling of call traces for exceptions of native functions
Due to the multiple prototypes being tried in chain, native functions
are throwing errors even in valid code, which will be swallowed when
finding a compatible prototype. In case of failure, the catched
exception message is sent back to the throwError API. This has 2
effects:

- due to the double usage of throwError, the details about the location
  and call trace get duplicated in the final message
- computing the call trace message is very costly due to the size of the
  call trace. Computing it for exceptions which will be swallowed is an
  useless performance hit.

Places catching exceptions are now toggling a flag which tells
throwError to skip adding the details. This avoids both the performance
hit of multiple prototypes and the duplicated details (as the first call
to throwError is now done with the flag toggled).