Refactor the handling of warnings in SassSpecTests
Instead of prepending them to the output and expectation to do a single
assertion, the assertion is now done separately. This makes it much less
confusing when the ouput is the right one but the warning is missing.
A new exclude list file is maintained, to cover tests for which the
compiler generates the right output but does not display the warning,
instead of skipping them entirely. This allows for more tests to run
against the spec (missing a warning is less an issue than getting the
output wrong). And this matches the behavior of libsass (the options.yml
in hrx files can declare warnings as not implemented by libsass).