Reorder color names to ensure that the same name is used than in libsass
Several colors have multiple names in CSS (including all gray colors).
The color map now ensures that the first name in the array is used, and
that colors are ordered in the same way than in libsass so that the
final name is the same.
Note that dart-sass uses a different rule for the canonical name, always
using the first one in alphabetical order (which matches libsass for
gray/grey but not for magenta/fuschia or cyan/aqua).