Improve test names for sass-spec by using a named provider
This makes it a lot easier to use phpunit --filter as using the file
name (without the hrx) exception now works rather than having to
discover the index.
The number of tests is back to 1-based in the names to make it less
confusion (tests now go again from 1/1980 to 1980 rather than from
0/1980 to 1979/1980), which is not an issue as phpunit --filter does not
require the usage of a 0-based index anymore.