Stop using includes to filter out module-based specs
Now that includes are shared for the whole hrx file, this filters out
too many specs. So the module filtering is now done only for the input
file. Specs which use modules in the includes but import them in the old
way in the input file are now handled by the exclude file instead.