Add phpstan in the project
For now, it complains about a lot of code, as it does not understand
everything about the legacy code in the compiler. But it will bring
benefits for new modern code. Thanks to the baseline feature, these can
be ignored
It can be run as "make phpstan" to analyse the code and report errors or
as "make phpstan-baseline" to regenerate the baseline (to remove fixed
errors or to ignore more errors in legacy code).