Fix the matching of the space ending an escape sequence
Unicode escape sequence can end with a space to delimit them from
the next char. This space is always allowed, and is necessary when
the sequence is shorter than 6 chars and the next char would be is
part of the hex alphabet (as it would be valid as part of the
sequence).