Remove useless use statements
Classes from the same namespace don't need to be imported.
Trying to use them to denote classes used in the file is flawed as PHP
does not require a use statement to be able to use a class and there is
nothing ensuring the list stays uptodate (tooling will flag them as
useless as it knows about the PHP behavior).