Fix #268 : add a checkImportResolutions cache option, that trigger a full check of unchanged findImport() before accepting a cache as valid The import resolution checking is stored in a class static property to avoid several identical checks which are involving a lot of disk access. This has a noticeable impact when compiling several files in a row, and using a framework and each file is using part of this framework. The bigger the framework the bigger the benefit :)
The compilation result also has an additional flag isCached set to true when the result is coming from the cache, and checkValid() will return true without any check if the result is not coming from the cache.
This allow the caller to call the checkValid() method without performance impact when the result has just been compiled.