Fix BC with projects disabling the CWD-based import resolution
When a path is not provided in compile(), scssphp 1.4.0 uses the current
working directory as the directory of the file, in an effort to keep BC
with the old behavior.
But this actually break things for projects which were removing that
import path. And when combining that with open_basedir restrictions, it
can create issues.