Introduce a LoggerInterface
The object-oriented interface makes it simpler to customize the
reporting of warnings (no need to use a special stream wrapper to keep
it realtime). It will also simplify future internal refactorings as
object references are easier to share than PHP resources.
As a nice benefit, the stderr stream opened automatically by the
compiler is now properly closed when it is not used anymore.