Require either mbstring or iconv
Sass strings are indexed in term of Unicode codepoints. Falling back to
byte-based indexes when neither mbstring nor iconv are available on the
system is not spec-compliant as it changes how string indexes are
working. iconv is enabled by default in PHP (this might not be true on
all distributions though as they customize how they build PHP) and
mbstring is quite common, especially now that UTF-8 is the norm.
Installing mbstring is recommended as it has better performance than
iconv. Scssphp is only using iconv as fallback when mbstring is not
available.