Refactor the signature for reduce
reduce() was returning null only when passing null as input. And the
implementation of native function calls was the only place relying on
passing null (due to the weird handling of named-only parameters relying
on passing null rather than a value). Handling null there directly
allows to remove null from the return type of reduce(), solving lots of
type checking issues in code passing a non-null value and relying on the
fact that the return type is not null.