Remove CSS imports from the includedFiles of the result object
a5f84108979d319158cc053e57bf957958ee37fd said that this was making it
consistent with the way dart-sass builds the list. But this was wrong,
due to a bad analysis of the dart-sass behavior. Code generating a CSS
import rule des not consider the URL of that import as an included file
for the sass rendering.