Refactor CompilationResult to be an immutable value object
Tracking the data during the compilation is now done again internally in
the Compiler. this gives more flexibility into refactoring the logic if
necessary. It also avoids storing the CompilationResult after the end of
the compilation for the getParsedFiles deprecated getter, which would
keep a reference to the generated CSS and sourcemap as well, potentially
increasing memory usage.
Data used only by our caching layer for freshness checks is now kept
outside the CompilationResult object, to avoid exposing it in the public
API.