Avoid changing the mbstring encoding when useless in the Parser
The Parser does not use mbstring itself. The only reason to change the
encoding is to ensure compatibility with the mbstring.func_overload
feature of PHP, as we rely on the fact that standard PHP functions are
dealing with bytes, not with multibyte chars (which would not be
compatible with preg_match offsets which are not overloaded).
There is no need to change that global state if that obsolete PHP
feature is not enabled.