Deprecate the scssphp-glob function
This function is not standard in Sass. It is totally untested. And it
has a confusing behavior: the argument must either be a glob based on an
absolute path (not easy to do in a Sass file if it must be portable) or
relative to the current working directory of the PHP process, not to the
Sass file using the function, (which is not in the control of the author
of the Sass file).