Fix an invalid test
This test breaks when trying to run it against dart-sass. The test
expects this to be a comma-separated list passed to the list-separator
function. But that's not what is actually done. The commas there are
separating function arguments. Defining a comma-separated list inline in
a function call requires using parenthesis.