Deprecate Compiler::compile in favor of Compiler::compileString
This restores backward compatibility on the compile() method by keeping
a string return value.
The modern API using a CompilationResult object is now available through
compileString instead. This modern API does not write the sourcemap to
the filesystem when configuring an external sourcemap. This is expected
to be done by the caller (similar to the fact the CSS is not written to
disk). However, the sourcemap comment is present in the CSS.