Refactor the handling of overloaded builtin functions
This now separate the selection of the candidate overload to apply
first, instead of trying to build arguments fully and catching exception
which might happen during that time (which was preventing giving more
precise errors for anything appearing inside a function call due to
catching any inner exception and reassigning them, in case we would
start tracking more precisely).
The selection of the overload to use matches the implementation of
dart-sass.