Use a * requirement for the sass-spec requirement
The sass-spec requirement refers to a package definition inline in our
composer.json. So we only ever have a single version available for the
solver. Using * in the requirement removes one step from the process to
update the pinned version of sass-spec.
Hardcoding a version of sass-spec and not changing it when updating the
reference is not an option. Composer would see that the vendor directory
already contains the same version that what is requested and would not
download the archive for the new commit.