Forbid extending complex selectors
Extending complex selectors is not supported in the official sass
implementation, and the spec does not define what that could mean.
The compiler was implementation that by ignoring parts of the selector,
keeping only the parent side of it. The result is something that has
absolutely no chance of matching what a developer could expect (and if
someone actually wanted that behavior, they can delete all the ignored
parts from their source code, making the code clear about what actually
gets extended).