Overhaul the variable registration API
The new API has 2 methods. One adding new variables (like the old
setVariables but with a name reflecting its behavior) and one replacing
all variables.
Values are expected to be converted to sass values using the new
ValueConverter API before being passed to the Compiler registration
methods (passing non-converted values triggers a deprecation warning).
This prepares for the refactoring of the representation of values
scheduled for a 2.x release, in which the ValueConverter will return
values using the new API. Because of that, the type returned by the
ValueConverter should be treated as opaque.