Refactor the implementation of sass argument lists
Argument lists are expected to be a list containing positional
arguments, with a way to access named arguments. The representation of
the value was mixing the named arguments with the positional ones in the
list, which was breaking the usage as a list and also returning wrong
data in the `keywords` function.
This is fixing bugs for the usage of an argument list in Sass code.
This is technically a BC break for custom host function that expect to
accept arbitrary named arguments while declaring arguments explicitly.
However, this is considered OK. This representation of argument lists is
not documented anywhere. And in previous releases of scssphp, there was
no such usages in the core either to allow discovering how to do it
(change-color, scale-color and adjust-color) were not implemented in
terms of a rest argument). The documentation about implementing custom
functions was showing how to achieve this for a function without an
argument declaration only. So it is unlikely that anyone discovered how
to do it in a way impacted by the BC break.