Rewrite the documentation
The documentation is now updated to cover the new APIs of the library.
Documentation about extension points of the library has been extracted
in a new section of the documentation and its content has been completed
to explain extension points better.

Links between pages have been replaced with relative links to the
markdown files (which are converted by the jekyll plugin), allowing
links to work fine when reading the markdown files on github directly
rather than using the website.