Remove the pscss CLI from the package
The CLI is still there in the repository as a maintenance tool allowing
to run the compiler to test it. However, the intended usage of the
package is to use the compiler programmatically. Projects able to rely
on a CLI tool to compile their Scss files should rather use the official
Sass CLI instead.