Fix the handling of charset rules in the Sass source
The output of the Sass compiler is always UTF-8. And a charset rule that
is not at the very beginning of the CSS file is useless anyway due to
how parsing CSS works.
The charset rule added in the output by the compiler depends only on the
presence of non-ASCII chars in the output.