Implement an option to skip the rendering of the charset rule
When using the output of the Sass compiler to combine it with other CSS
or to put it in an inline style tag, the charset rule is useless as it
will be ignored by the CSS parser. This option allows skipping it.
This option is the exact equivalent of the `charset` option of the
dart-sass API.