Add a deprecation warning when calling core functions with wrong names
Due to the internal implementation of the function discovery for core
functions, extra names were available for all functions with different
positions of dashes/underscores in the function name (additional ones or
omitted ones) or a different case. Those functions names are not
standard and would be treated as CSS functions by official Sass
implementations.