Add tests against the Bourbon framework
This framework is also tested by the new CI jobs of dart-sass, alongside
Foundation, Bootstrap 4 and Bootstrap 5 (that we already test) and Bulma
(which uses the indented syntax and so not supported by scssphp).
As bourbon is a Sass framework providing mixins and functions rather
than a CSS framework, we test it by compiling their test fixtures (as
done by dart-sass).