Make the comparison script able to use sassc
The sass executable being used is now configurable through the
`SASS_EXECUTABLE` environment variable, allowing to use `sassc` instead
of `sass` for instance.
The options used when calling the executable are also restricted to
options common to sassc and dart-sass (so no `--no-source-map` option,
which was useless anyway as that's the default in dart-sass when writing
the CSS to stdout) and the expanded style is configured explicitly (as
that's not the default in libsass).