Fix php8.1 strpos deprecated message for pure css imports
In case you import a css file `$filename` is null. That causes the following warning:

```
Deprecated: strpos(): Passing null to parameter #1 ($haystack) of type string is deprecated in /var/www/html/vendor/scssphp/scssphp/src/SourceMap/SourceMapGenerator.php on line 355
```

By adding a check if `$filename` is set (and not null), the provided null-value to strpos is avoided.