Upgrade phpstan
As the work on 2.0 is not done yet and some activity still happens on
the 1.x branch, it makes sense to use an uptodate version of phpstan.

An exact version is used to avoid random CI failure caused by phpstan
updates that might change the message excluded by the baseline.