Split invalid SCSS inputs into their own input tests
Our input tests are including a few behaviors that are considered errors
in dart-sass (color arithmetic and broken interpolation in loud
comments). Those behavior are deprecated in scssphp but still supported
for BC.
Extracting those into dedicated inputs for the testsuite allows
comparing the other parts of those tests to the official implementation.