Rework the late initialization of the MixinRule::$content property
PHP static analyzers don't support properly the pattern of late
initialization right now.
The new code defines the property as nullable and not read-only, to
allow changing the property later in case it is null. This is less
explicit about the intent of the property, but it is supported by the
tooling.