Add an API for location-aware loggers
This new interface will give more flexibility to loggers to report
locations in a structured way.
Once the rewrite is complete, all internal usages of the logger will
work with that interface, letting the logger deal with adding the
location in the message for textual output.
The AdaptingLogger provides the compatibility layer to support loggers
implemented against the 1.x interface.