Assert that `$n` is an integer for `nth($list, $n)` and `set-nth($list, $n, $value)`
In PHP 8.1 implicit conversions from floats to integers trigger
a deprecation warning.
The change from `assertNumber()` to `assertInteger()` falls in
line with the behavior of Dart Sass, which throws an exception
if the index isn't a number, if that number isn't an integer,
or if that integer isn't a valid index for the given list.