Emit a deprecation warning for the strict parsing of unary operators
This emits a deprecation warning for `$a -$b` and `$a +$b`, since
these look like they could be unary operations but they're actually
parsed as binary operations. Either explicitly write `$a - $b` or
`$a (-$b)`. See https://sass-lang.com/d/strict-unary for more details.

This deprecation is only implemented for the new parser ported from
dart-sass, not for the legacy parser.