Enable bleeding edge rules for phpstan
As the new code works a lot with the list type, we should make phpstan
enforce it, which is the case only in bleeding edge for now.
As phpstan is pinned to an exact version, the fact that bleeding edge
can enable new rules in any release (and so change the result of the
analysis) is not an issue.