Fix the compilation cache key when using numbers in variables
Variables gets encoded as JSON as part of the cache key. For most types,
this is fine as they are represented as arrays. But numbers are
represented as an object with private properties so all numbers were
encoded the same, breaking the cache invalidation when replacing them
with a different number.