Add an explicit dependency on the mbstring polyfill
The polyfill is already a transitive dependency because of
`symfony/filesystem`. And all the code of the new API is already
implemented by using mbstring directly (assuming it is either available
or polyfilled).
The ponyfill utilities used in the 1.x codebase have been simplified
internally but not replaced as the code using them will be deleted soon
so it was not worth it.