Fix the detection of trailing comments
The behavior of the offset in `strrpos` does not match the behavior of
the offset in Dart's `String.lastIndexOf`. Instead, positive offsets are
skipping the part of the string before that offset rather than after
that offset. To replicate the behavior of Dart, we need to use a
negative offset (counted from the end of the string).