Fix compatibility with PHP 8.1
The combination of first-class callables and Exceptions may cause a crash in
PHP 8.1 that is fixed for PHP 8.2+ in
php/php-src@b3e26c3036a54e9821ea7119c26cdabe484fe36d.

It specifically can be triggered by the use of `SaasNumber::isComparableTo()`
for incomparable numbers and is caused by the
`NumberUtil::fuzzyGreaterThan(...)` FCC as used in `SaasNumber::greaterThan()`.

Fix this issue by calling `$other->coerceValueToMatch($this);` directly, which
is the function that is responsible for throwing the `SassScriptException`
inside of `SaasNumber::greaterThan()`.

see scssphp/scssphp#746
see scssphp/scssphp#751
see https://github.com/scssphp/scssphp/issues/752#issuecomment-2423857568