Remove check redundant with static analysis
dart-sass performs some checks for the types of elements in an
Interpolation because Dart does not support union types (and so this is
not checked by the compiler).
This check is redundant in our case because we use a union type for the
precise type, and phpstan detects this as dead code due to the condition
always being false.