Add a `--trace` option in our CLI
This option (inspired by the equivalent option of dart-sass) will
display the PHP stack trace for any SassException reported during the
execution, to make it easier to debug issues when we don't actually
expect to get this exception.

This is not enabled by default to keep our (contributor-only) CLI in
sync with the expectations of the upstream sass-spec runner, to be able
to run the official sass-spec through the official runner if needed (not
yet working due to our custom skipping of module-based tests).