Remove additional empty line added by our fake HRW parsing
The way we (don't) parse HRX file in our runner was injecting an extra
LF at the end of each file (except maybe the last one).
This can have an impact when that file is the input file (which is never
the last file). Output content is always trimmed at the end before
comparison so this removal won't be an issue for the last file.