Implement the `if` function for usage with `get-function` and `call`
Direct usages of the `if()` function are parsed as a special
`IfExpression` to allow short-circuiting arguments.
The case of the indirect usage of the function is only covered by an
error spec in the sass-spec testsuite, which was hiding the fact that we
missed implementing it (as we still got an error, but with a different
message that we were not asserting).