Fix the serialization of values when using them outside the compiler
The contravariant rules related to the `CssVisitor` and the
`ModifiableCssNode` layers trigger a bug in PHP where some loading
orders for classes are working while others fail at resolving the
classes properly, saying they don't exist.

The Compiler was already applying a workaround for this issue before
using the AST classes. However, the `Value` API relies on the
`SerializeVisitor` for its string representations. When using values
outside the compiler (for instance when testing custom functions in unit
tests), the same issue could be triggered. The workaround is now also
applied in the Serializer for those usages.