Ensure that the URI resolution semantic is consistent with Dart
The dart `Uri.resolve` API implements 3 behaviors that differ from
the RFC3986 resolution algorithm:
1. when resolving a relative path reference against a base URI with a
  scheme, no authority and a relative path, it treats it as if the base
  URI path was absolute (which seems to violate the behavior of the RFC)
2. it supports resolving a relative path reference against a relative
  path base URL (instead of being an error case in the RFC3986) by
  merging the relative paths together to produce a new relative path
  (which preserves leading `..` segments, unlike usual path resolution
  on absolute URLs)
3. it supports resolving any kind of reference against a base URI without
  a scheme (for cases other than the previous case) by applying the
  algorithm of RFC3986 with an undefined scheme instead of a defined
  scheme (producing an undefined scheme in the result).

Case 1 and 2 are now implemented in our `UriUtil` to match the Dart
behavior. The case 3 is covered already by the logic available in
`league/uri` 7.5, which is why this inconsistency was not discovered
until now.
Making that logic compatible with `league/uri` 7.6 (which performs
strict validation of the RFC3986 algorithm) will be done separately.