Upgrade to league/uri 7.6
league/uri 7.6 is stricter in its validation logic for the URI
resolution algorithm to enforce behavior of the RFC3986.
This implements the other cases supported by Dart's URI resolution
algorithm in our own utility.
This bumps the min version of league/uri to version 7.6 as supporting
both versions 7.5 and 7.6 in this new logic is harder.